@@ -1,18 +1,16 @@
-// src/components/Navbar.js
+// Navbar.js
 import React from 'react';
 import { Link } from 'react-router-dom';
 
 const Navbar = () => {
   return (
-    <nav>
-      <ul>
-        <li><Link to="/mood-tracker">Mood Tracker</Link></li>
-        <li><Link to="/song-playlist">Song Playlist</Link></li>
-        <li><Link to="/would-you-rather">Would You Rather</Link></li>
-        <li><Link to="/daily-quote">Daily Quote</Link></li>
-        <li><Link to="/password-vault">Password Vault</Link></li>
-      </ul>
-    </nav>
+    <div className="navbar">
+      <Link to="/mood-tracker">Mood Tracker</Link>
+      <Link to="/song-playlist">Song Playlist</Link>
+      <Link to="/would-you-rather">Would You Rather</Link>
+      <Link to="/daily-quote">Daily Quote</Link>
+      <Link to="/password-vault">Password Vault</Link>
+    </div>
   );
 };
 
