@@ -1,11 +1,14 @@
 import React, { useState, useEffect } from 'react';
-import axios from 'axios';
+import axios from 'axios';  // Import axios here
 
 const DailyQuote = () => {
   const [quote, setQuote] = useState('');
 
   useEffect(() => {
-    axios.get('https://api.quotable.io/random').then((response) => setQuote(response.data.content));
+    // Fetch the quote when the component mounts
+    axios.get('https://api.quotable.io/random')
+      .then((response) => setQuote(response.data.content))
+      .catch((error) => console.error('Error fetching quote:', error));
   }, []);
 
   return (
@@ -17,4 +20,3 @@ const DailyQuote = () => {
 };
 
 export default DailyQuote;
-
